[IMP] Taxes: fix tax computation / included in price
task-5094417

At the moment, the section for the `Percentage of Price tax included`
tax computation doesn't distinguish itself enough from the 'Included in
Price' option which people usually want.

Percentage of Price Tax Included taxes are very rare and only used in
countries where tax rates are expressed as a percentage of the
tax-included total.

We also take the opportunity to improve the description of the
`Percentage of Price` computation type and the `Included in Price`
option.